Refactor services into Interfaces and Impl
