+++ project-management-backend/src/main/java/com/biswas/project_management_backend/controller/CompanyController.java
@@ -2,16 +2,17 @@ package com.biswas.project_management_backend.controller;
 
 import com.biswas.project_management_backend.dto.CompanyDto;
 import com.biswas.project_management_backend.service.CompanyService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/api/company")
+@RequiredArgsConstructor
 public class CompanyController {
 
-    @Autowired
-    CompanyService companyService;
+    private final CompanyService companyService;
 
     @GetMapping("/{companyId}")
     public ResponseEntity<CompanyDto> getCompanyById(@PathVariable Long companyId) {
