+++ project-management-backend/src/main/java/com/biswas/project_management_backend/controller/UserController.java
@@ -20,9 +20,7 @@ import java.util.Set;
 public class UserController {
 
     private final UserService userService;
-
-    @Autowired
-    private AnalyticsService analyticsService;
+    private final AnalyticsService analyticsService;
 
     @GetMapping("/company/{companyId}")
     public List<UserDto> getAllUsers(@PathVariable Long companyId){
