+++ project-management-backend/src/main/java/com/biswas/project_management_backend/service/AnalyticsService.java
@@ -19,74 +19,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-@Service
-public class AnalyticsService {
-
-    @Autowired
-    private TaskRepository taskRepository;
-
-    @Autowired
-    private ProjectRepository projectRepository;
-
-    @Autowired
-    private TeamRepository teamRepository;
-
-    @Autowired
-    private UserRepository userRepository;
-
-    public AdminAnalyticsDto getAnalyticsSummary(Long companyId, String dateFrom, String dateTo) {
-        LocalDate from = dateFrom != null ? LocalDate.parse(dateFrom) : null;
-        LocalDate to = dateTo != null ? LocalDate.parse(dateTo) : LocalDate.now();
-        AdminAnalyticsDto dto = new AdminAnalyticsDto();
-        dto.setTotalUsers(userRepository.countByCompanyId(companyId));
-        dto.setTotalProjects(projectRepository.countByCompanyId(companyId));
-        dto.setTotalTeams(teamRepository.countByCompanyId(companyId));
-        dto.setTotalTasks(taskRepository.countByCompanyId(companyId));
-        Map<TaskStatus, Long> tasksByStatus = new EnumMap<>(TaskStatus.class);
-        for (TaskStatus status : TaskStatus.values()) {
-            tasksByStatus.put(status, taskRepository.countByStatusAndCompanyId(status, companyId));
-        }
-        dto.setTasksByStatus(tasksByStatus);
-        Map<ProjectStatus, Long> projectsByStatus = new EnumMap<>(ProjectStatus.class);
-        for (ProjectStatus status : ProjectStatus.values()) {
-            projectsByStatus.put(status, projectRepository.countByStatusAndCompanyId(status, companyId));
-        }
-        dto.setProjectsByStatus(projectsByStatus);
-        dto.setOverdueTasks(taskRepository.countByDueDateBeforeAndStatusNotAndCompanyId(LocalDate.now(), TaskStatus.DONE, companyId));
-        dto.setActiveUsersLastWeek(userRepository.countActiveUsersLastWeek(companyId, to.minusDays(7)));
-        return dto;
-    }
-
-
-    public UserAnalyticsDto getUserAnalytics(String userEmail, Long companyId) {
-        User user = userRepository.findByEmailAndCompanyId(userEmail, companyId);
-//                .orElseThrow(() -> new RuntimeException("User not found"));
-
-        Long assignedTasks = taskRepository.countByAssignee(user);
-        Long completedTasks = taskRepository.countByAssigneeAndStatus(user, TaskStatus.DONE);
-        Long overdueTasks = taskRepository.countOverdueTasksByAssignee(user.getId());
-
-        List<Object[]> results = taskRepository.countTasksByStatusForUser(user.getId());
-        Map<TaskStatus, Long> userTasksByStatus = results.stream()
-                .collect(Collectors.toMap(
-                        row -> (TaskStatus) row[0],
-                        row -> (Long) row[1]
-                ));
-
-        Long totalProjects = projectRepository.countByMembersContaining(user);
-        Long activeProjects = projectRepository.countByMembersContainingAndStatus(user, ProjectStatus.ACTIVE);
-        Long completedProjects = projectRepository.countByMembersContainingAndStatus(user, ProjectStatus.COMPLETED);
-        Long totalTeams = teamRepository.countByMembersContaining(user);
-
-        return new UserAnalyticsDto(
-                assignedTasks,
-                completedTasks,
-                overdueTasks,
-                userTasksByStatus,
-                totalProjects,
-                activeProjects,
-                completedProjects,
-                totalTeams
-        );
-    }
+public interface AnalyticsService {
+    AdminAnalyticsDto getAnalyticsSummary(Long companyId, String dateFrom, String dateTo);
+    UserAnalyticsDto getUserAnalytics(String userEmail, Long companyId);
 }
