+++ project-management-backend/src/main/java/com/biswas/project_management_backend/service/AuthService.java
@@ -1,40 +1,6 @@
 package com.biswas.project_management_backend.service;
 
-import com.biswas.project_management_backend.model.User;
-import com.biswas.project_management_backend.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Service;
+public interface AuthService {
 
-import java.util.Optional;
-
-@Service
-@RequiredArgsConstructor
-public class AuthService {
-
-    @Autowired
-    UserRepository userRepository;
-
-    @Autowired
-    PasswordEncoder passwordEncoder;
-
-    public boolean changePassword(String username, String oldPassword, String newPassword) {
-        Optional<User> optionalUser = userRepository.findByUsername(username);
-
-        if (optionalUser.isEmpty()) {
-            return false;
-        }
-
-        User user = optionalUser.get();
-
-        if (!passwordEncoder.matches(oldPassword, user.getPassword())) {
-            return false;
-        }
-
-        user.setPassword(passwordEncoder.encode(newPassword));
-        userRepository.save(user);
-
-        return true;
-    }
+    boolean changePassword(String username, String oldPassword, String newPassword);
 }
